Refactor register parachain
