+++ ts-tests/config.ci.json
@@ -3,5 +3,8 @@
     "private_key" : "0xe82c0c4259710bb0d6cf9f9e8d0ad73419c1278a14d375e5ca691e7618103011",
     "ocw_account" : "5FEYX9NES9mAJt1Xg4WebmHWywxyeGQK8G3oEBXtyfZrRePX",
     "parachain_ws" : "ws://localhost:9954",
-    "relaynode_ws" : "ws://localhost:9944"
+    "relaynode_ws" : "ws://localhost:9944",
+    "parachain_id": 2022,
+    "parachain_wasm_path": "para-2022-wasm",
+    "parachain_genesis_path": "para-2022-genesis"
 }
